feat: add `/api/last` and `/api/week` endpoints
`/api/last` endpoint returns the latest temperature data for a given
station id.

`/api/week` endpoint returns an array of temperature observations for a
given station id. For readability, not all observations are returned.
Instead, only the high and low for any day, and only data from the last
week is returned.